+++ src/ui.cpp
@@ -1,48 +1,8 @@
-#include "helper.h"
+#include "ui.h"
 #include <iostream>
 #include <cstdlib>
 #include <vector>
 #include "raylib.h"
-
-
-INT_MATRIX_2D::INT_MATRIX_2D(int rows, int columns) {
-	this->rows = rows;
-	this->columns = columns;
-
-	int **array = (int **)(malloc(sizeof(int *) * rows));
-
-	for (int i = 0; i < rows; i++) {
-		array[i] = (int *)malloc(sizeof(int) * columns);
-
-		for (int x = 0; x < columns; x++ ) {
-			array[i][x] = 0;
-
-		}
-	}
-    
-
-	this->array = array;
-}
-
-int* INT_MATRIX_2D::operator[](int i) {
-	if (i > this->rows) {
-		std::cout << i <<" is out of bountsds";
-		exit(-1);
-	}
-	int *bruh = this->array[i];
-
-	return bruh;
-}
-
-INT_MATRIX_2D::~INT_MATRIX_2D() {
-	for (int i = 0; i < this->rows; i++) {
-		free(this->array[i]);
-	}
-	free(this->array);
-}
-
-
-
 
 void WriteToMatrix(INT_MATRIX_2D &matrix, int downscale_amt, int xResolution, int yResolution, std::vector<POINT_OF_INTREST> &pois) {
     Vector2 mousePosition = GetMousePosition();
@@ -57,7 +17,7 @@ void WriteToMatrix(INT_MATRIX_2D &matrix, int downscale_amt, int xResolution, in
             matrix[xMousePosition][yMousePosition] = 0;
 
         } else if (IsKeyPressed(KEY_H)) {
-            pois.push_back(POINT_OF_INTREST(xMousePosition,yMousePosition,std::string("ID"),false,"NAME"));
+            pois.push_back(POINT_OF_INTREST(xMousePosition,yMousePosition,std::string("ID"),false));
             matrix[xMousePosition][yMousePosition] = 2;
         }
 
@@ -78,30 +38,23 @@ bool IsPointOfInterest(int xPosition, int yPosition, std::vector<POINT_OF_INTRES
     return false;
 }
 
-POINT_OF_INTREST::POINT_OF_INTREST(int xPosition, int yPosition, std::string uuid,bool arduino, std::string name) {
-	this->xPosition = xPosition;
-	this->yPosition = yPosition;
-	this->uuid = uuid;
-    this->arduino = arduino;
-	this->name = name;
-}
 
-POI_RECTANGLE_UI::POI_RECTANGLE_UI(Rectangle positionRectangle, Rectangle idRectangle, Rectangle nameRectangle,Rectangle toggleArduinoRectangle, Rectangle deleteRectangle, POINT_OF_INTREST *poi) {
+POI_RECTANGLE_UI::POI_RECTANGLE_UI(Rectangle positionRectangle, Rectangle idRectangle,Rectangle toggleArduinoRectangle, Rectangle deleteRectangle, POINT_OF_INTREST *poi) {
     this->positionRectangle = positionRectangle;
-    this->nameRectangle = nameRectangle;
     this->idRectangle = idRectangle;
     this->toggleArduinoRectangle = toggleArduinoRectangle;
     this->deleteRectangle = deleteRectangle;
     this->poi = poi;
 }
+
+
 void POI_RECTANGLE_UI::Draw() {
     Rectangle idRectangle = this->idRectangle;
-    Rectangle nameRectangle = this->nameRectangle;
+
     Rectangle positionRectangle = this->positionRectangle;
     Rectangle arduinoToggleRectangle = this->toggleArduinoRectangle;
     Rectangle deleteRectangle = this->deleteRectangle;
     DrawRectangleRec(positionRectangle,GREEN);
-    DrawRectangleRec(nameRectangle,GREEN);
     DrawRectangleRec(idRectangle,GREEN);
     DrawRectangleRec(deleteRectangle, RED);
     
@@ -119,7 +72,7 @@ void POI_RECTANGLE_UI::Draw() {
     combined.append(yPositionStr);
 
     DrawText(combined.c_str(),positionRectangle.x+3,positionRectangle.y+2,13,WHITE);
-    DrawText(poi->name.c_str(),nameRectangle.x+3,nameRectangle.y+2,13,WHITE);
+    
     DrawText(poi->uuid.c_str(),idRectangle.x+3,idRectangle.y+2,13,WHITE);
     DrawText("X",deleteRectangle.x+7, deleteRectangle.y+2,13,WHITE);
 
@@ -146,13 +99,12 @@ void DrawUI(std::vector<POINT_OF_INTREST> &pois, int xResolution, int yResolutio
     for (int i = 0; i < pois.size(); i++) {
         int yPosition = startingY+(20*i);
         Rectangle positionRectangle = {startingX,yPosition,45,15};
-        Rectangle idRectangle = {startingX + 50, yPosition,100,15};
-        Rectangle toggleArduinoRectangle = {startingX + 50 + 100 + 5,yPosition, 25, 15};
-        Rectangle nameRectangle = {startingX + 50 + 100 + 5 + 25 +5 , yPosition, 75, 15};
-        Rectangle deleteRectangle = {startingX + 50 + 100 + 5 + 25 + 5 + 75 + 5, yPosition, 25, 15};
+        Rectangle idRectangle = {startingX + 50, yPosition,175,15};
+        Rectangle toggleArduinoRectangle = {startingX + 50 + 175 + 5,yPosition, 25, 15};
+        Rectangle deleteRectangle = {startingX + 50 + 175 + 5 + 25 + 5, yPosition, 25, 15};
         POINT_OF_INTREST *poi = &pois[i];
 
-        POI_RECTANGLE_UI poiRectangle(positionRectangle,idRectangle,nameRectangle,toggleArduinoRectangle,deleteRectangle,poi);
+        POI_RECTANGLE_UI poiRectangle(positionRectangle,idRectangle,toggleArduinoRectangle,deleteRectangle,poi);
         poiRectangle.Draw();
         uiRectangles.push_back(poiRectangle);
     }
@@ -164,16 +116,13 @@ void GetUserInput(std::vector<POI_RECTANGLE_UI> &pois, std::vector<POINT_OF_INTR
         POI_RECTANGLE_UI poi = pois[i];
         
         Rectangle arduinoToggleRec = poi.toggleArduinoRectangle;
-        Rectangle nameRectangle = poi.nameRectangle;
         Rectangle uuidRectangle = poi.idRectangle;
         Rectangle deleteRectangle = poi.deleteRectangle;
         if (CheckCollisionPointRec(GetMousePosition(),arduinoToggleRec) && IsMouseButtonPressed(MOUSE_BUTTON_LEFT)) {
             poi.poi->arduino = !poi.poi->arduino;
         }
 
-        if (CheckCollisionPointRec(GetMousePosition(),nameRectangle)) {
-            GetStringFromUser(poi.poi->name);            
-        }
+
         if (CheckCollisionPointRec(GetMousePosition(),uuidRectangle)) {
             GetStringFromUser(poi.poi->uuid);
         }
@@ -193,7 +142,6 @@ void GetUserInput(std::vector<POI_RECTANGLE_UI> &pois, std::vector<POINT_OF_INTR
 }
 
 void GetStringFromUser(std::string &str) {
-    
     int key = GetCharPressed();
 
     while (key > 0 && str.length() < 9) {
